@@ -26,8 +26,9 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.PowerManager;
 import android.os.SystemProperties;
-import android.support.v4.app.NotificationCompat;
-import android.support.v7.preference.PreferenceManager;
+
+import androidx.core.app.NotificationCompat;
+import androidx.preference.PreferenceManager;
 
 import org.evolutionx.ota.misc.Constants;
 import org.evolutionx.ota.misc.Utils;
