@@ -17,7 +17,7 @@
  */
 package org.evolutionx.ota;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 public abstract class UpdatesListActivity extends AppCompatActivity {
     public abstract void showSnackbar(int stringId, int duration);
